+++ src/main/java/io/vn/dungxnd/duckabary/core/Utils.java
@@ -8,4 +8,30 @@ public class Utils {
         DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
         return time.format(formatter);
     }
+
+    public static LocalDateTime getDateTimeFromString(String time) {
+        if (time.contains("(") && time.contains(")")) {
+            time = time.substring(time.indexOf("(") + 1, time.indexOf(")"));
+        }
+
+        if (time.length()==16){
+            time = time + ":00";
+        }
+
+        if (time.length()==13){
+            time = time + ":00:00";
+        }
+
+        DateTimeFormatter formatter = DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm:ss");
+        return LocalDateTime.parse(time, formatter);
+    }
+
+    public static boolean isValidDateTime(String time) {
+        try {
+            getDateTimeFromString(time);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
 }