+++ src/main/java/io/vn/dungxnd/duckabary/core/library_management/LibraryCmdService.java
@@ -1,6 +1,5 @@
 package io.vn.dungxnd.duckabary.core.library_management;
 
-import io.vn.dungxnd.duckabary.core.user_management.User;
 import io.vn.dungxnd.duckabary.core.user_management.UserCmdService;
 
 import java.util.ArrayList;
@@ -76,48 +75,6 @@ public class LibraryCmdService extends LibraryService {
         return documents;
     }
 
-    @Override
-    public boolean borrowDocumentByID(int userId, int docId) {
-        if (super.getUserByID(userId) == null) {
-            System.out.println("User with id " + userId + " not found");
-            return false;
-        }
-        if (super.getDocumentByID(docId) == null) {
-            System.out.println("Document with id " + docId + " not found");
-            return false;
-        }
-        if (!super.borrowDocumentByID(userId, docId)) {
-            System.out.println("Document with id " + docId + " is out of stock");
-            return false;
-        }
-        System.out.printf(
-                "Document '%s' with id %d borrowed by user %s (%d) successfully\n",
-                getDocumentByID(docId).getTitle(),
-                docId,
-                getUserByID(userId).getUsername(),
-                userId);
-        return true;
-    }
-
-    @Override
-    public boolean returnDocumentByID(int userId, int docId) {
-        if (super.getDocumentByID(docId) == null) {
-            System.out.println("Document with id " + docId + " not found");
-            return false;
-        }
-        if (!super.returnDocumentByID(userId, docId)) {
-            System.out.println(
-                    "Document with id " + docId + " is not borrowed by user with id " + userId);
-            return false;
-        }
-        System.out.println(
-                "Document "
-                        + getDocumentByID(docId).getTitle()
-                        + " with id "
-                        + docId
-                        + " returned successfully");
-        return true;
-    }
 
     @Override
     public UserCmdService getUserService() {